refactor(capture): extract authoring record builders
